Disable certain pages in PDF when data is missing
